Struct map::gpio::AfrhR

#[repr(C)]

pub struct AfrhR { /* fields omitted */ }

Methods

impl AfrhR

fn afrh15(&self) -> u8

Bits 28:31 - Alternate function selection for port x bit y (y = 8..15)

fn afrh14(&self) -> u8

Bits 24:27 - Alternate function selection for port x bit y (y = 8..15)

fn afrh13(&self) -> u8

Bits 20:23 - Alternate function selection for port x bit y (y = 8..15)

fn afrh12(&self) -> u8

Bits 16:19 - Alternate function selection for port x bit y (y = 8..15)

fn afrh11(&self) -> u8

Bits 12:15 - Alternate function selection for port x bit y (y = 8..15)

fn afrh10(&self) -> u8

Bits 8:11 - Alternate function selection for port x bit y (y = 8..15)

fn afrh9(&self) -> u8

Bits 4:7 - Alternate function selection for port x bit y (y = 8..15)

fn afrh8(&self) -> u8

Bits 0:3 - Alternate function selection for port x bit y (y = 8..15)