Struct map::gpio::AfrlW

#[repr(C)]

pub struct AfrlW { /* fields omitted */ }

Methods

impl AfrlW

fn reset_value() -> Self

Reset value

fn afrl7(&mut self, value: u8) -> &mut Self

Bits 28:31 - Alternate function selection for port x bit y (y = 0..7)

fn afrl6(&mut self, value: u8) -> &mut Self

Bits 24:27 - Alternate function selection for port x bit y (y = 0..7)

fn afrl5(&mut self, value: u8) -> &mut Self

Bits 20:23 - Alternate function selection for port x bit y (y = 0..7)

fn afrl4(&mut self, value: u8) -> &mut Self

Bits 16:19 - Alternate function selection for port x bit y (y = 0..7)

fn afrl3(&mut self, value: u8) -> &mut Self

Bits 12:15 - Alternate function selection for port x bit y (y = 0..7)

fn afrl2(&mut self, value: u8) -> &mut Self

Bits 8:11 - Alternate function selection for port x bit y (y = 0..7)

fn afrl1(&mut self, value: u8) -> &mut Self

Bits 4:7 - Alternate function selection for port x bit y (y = 0..7)

fn afrl0(&mut self, value: u8) -> &mut Self

Bits 0:3 - Alternate function selection for port x bit y (y = 0..7)