Struct map::views::map_controller_view::MapControllerView

pub struct MapControllerView<'a, T: MasterHal> { /* private fields */ }

allows you to control the map

Implementations

impl<'a, T: MasterHal> MapControllerView<'a, T>

pub fn new(map: Rc<RefCell<Map<MapStateInitialized>>>, hal: &'a T) -> Self

Creates a new MapControllerView

pub fn set_train_speed(&self, train: Train, speed: i8) -> Result<()>

Sets the speed of a train

Errors

Returns an error if the connection to the train is lost Returns an error if the train does not exist

pub fn stop_train(&self, train: Train) -> Result<()>

Stops a train by setting the speed to 0

Errors

Returns an error if the connection to the train is lost Returns an error if the train does not exist

pub fn set_switch(&self, switch: Switch, position: SwitchPosition) -> Result<()>

Sets the position of a switch

Errors

Returns an error if the connection to the switch is lost Returns an error if the switch does not exist

pub fn set_switch_straight(&self, switch: Switch) -> Result<()>

Sets the position of a switch to straight

Errors

Returns an error if the connection to the switch is lost Returns an error if the switch does not exist

pub fn set_switch_diverted(&self, switch: Switch) -> Result<()>

Sets the position of a switch to diverted

Errors

Returns an error if the connection to the switch is lost Returns an error if the switch does not exist

pub fn set_switch_between(&self, p1: Position, p2: Position) -> Result<()>

Sets the switch between two nodes in a way that the train can pass between them

args:

p1: Position of the first node

p2: Position of the second node

Errors

Returns an error if the nodes do not exist

Returns an error if there is no link between the two nodes

Returns an error if the connection to the switch is lost

Nodes

if there is no switch between the two nodes, this function does nothing

pub fn move_train(&self, train: Train, position: Position) -> Result<()>

Move a train to a position

Errors

Returns an error if the train does not exist

Returns an error if the node does not exist

Returns an error if the node is occupied or locked by another train

Returns an error if the position is not adjacent to the current position of the train

notes

if the train is already in the position, this function does nothing

this function dose not move the train “in the real world”, only in the map

the function will unlock the node that the train was in before

pub fn lock_node(&self, position: Position, train: Train) -> Result<()>

Locks a node for a train

Errors

Returns an error if the node does not exist

Returns an error if the train does not exist

Returns an error if the node is already locked/occupied by another train

pub fn unlock_node(&self, position: Position) -> Result<()>

Unlocks a node

Errors

Returns an error if the node does not exist

Returns an error if the node is not locked by any train

Returns an error if the node is occupied by a train

pub fn get_speed_to_reach(&self, train: Train, node: Position) -> Result<i8>

Gets the speed that a train needs to reach a node

the function keeps track of the direction of the train, as well as the max speed of the link between the two nodes

Errors

Returns an error if the train does not exist

Returns an error if the node does not exist

Returns an error if the node is not adjacent to the current position of the train

pub fn set_train_status( &mut self, train: Train, status: TrainStatus ) -> Result<()>

Sets the status of a train

Errors

Returns an error if the train does not exist

pub unsafe fn move_train_unchecked( &mut self, train: Train, position: Position ) -> Result<()>

move the train to a position this function is unsafe because it does not check if the position can be reached it also don’t update the direction of the train, to this function must be always reverted after been used