Struct map::views::map_creation_view::MapCreationView

pub struct MapCreationView { /* private fields */ }

This struct is used to create a map

Implementations

impl MapCreationView

pub fn new() -> Self

Create a new empty map

pub fn to_json(&self) -> Result<String>

pub fn to_map(self) -> Map<MapStateUninitialized>

return an uninitialized map

pub fn add_node(&mut self, node: Position) -> Result<()>

add one node to the map

errors

can returns an error if the node already exists

pub fn add_nodes(&mut self, nodes: &[Position]) -> Result<()>

add a list of nodes to the map

errors

can returns an error if at least one of the nodes already exists

pub fn add_switch(&mut self, switch: Switch) -> Result<()>

add a switch to the map

errors:

can return an error if the switch already exists

pub fn add_switches(&mut self, switches: &[Switch]) -> Result<()>

add a list of switches to the map

errors:

can return an error if at least one of the switches already exists

pub fn add_switch_station( &mut self, switch: Switch, position_center: Position, position_diverted: Position, position_straight: Position ) -> Result<()>

add switch station, aka 2 links between 3 nodes with the switch

arguments:

switch: the id of the switch

position_center: the id of the node that can be reached from the two other nodes

position_diverted: the id of the node that can be reached from position_center when the switch is diverted

position_straight: the id of the node that can be reached from position_center when the switch is straight

errors:

can return an error if the switch does not exist

can return an error if at least one of the nodes does not exist

can return an error if one of the nodes is already linked to something else

can return an error if the switch is already used

can return an error if at least one link is already been added to the map

can return an error if at least one train is already been added to the map

pub fn add_link( &mut self, n1: Position, n2: Position, max_speed_n1_to_n2: i8, max_speed_n2_to_n1: i8, length: u32 ) -> Result<()>

add a link between two nodes

arguments:

from: the id of the node where the link starts

to: the id of the node where the link ends

max_speed: the maximum speed allowed on the link

length: the length of the link

errors:

can return an error if the link already exists, or the node is already linked to something else

can return an error if at least one train has already been added to the map

pub fn add_train( &mut self, train: Train, position: Position, pointing_to: Option<Position> ) -> Result<()>

add one train to the map

arguments:

train: the id train to add

position: the id of the node where the train is located

pointing_to: the id of the node where the train is pointing to, this means that if pointing_to is P2, for examole, the train will reach P2 if the speed is positive if pointing_to is None, the train will point towards the dead end of the link, and will derail if the speed is positive

errors:

can return an error if the train already exists

can return an error if the node does not exist

can return an error if the node is already occupied by another train

can return an error if pointing_to is not None and the node is not in the adjacent nodes

can return an error if pointing_to is None and the node does not have a dead end